% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{auto.aspect}
\alias{auto.aspect}
\title{Choose aspect ratio}
\usage{
auto.aspect(x, y)
}
\arguments{
\item{x, y}{numeric vectors of coordinates defining a continuous curve, or
multiple curves delimited by \code{NA}s.  Alternatively, \code{x} can be a
list of two vectors.}
}
\value{
An aspect ratio, suitable for the \code{asp} parameter to
\code{\link{plot}} or \code{\link{plot.window}}.
}
\description{
Uses Cleveland's "banking to 45" rule to determine an optimal aspect ratio
for a plot.
}
\examples{

data(lynx)
plot(lynx)
plot(lynx,asp=auto.aspect(time(lynx),lynx))

}
\author{
Tom Minka
}
