% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{bhist}
\alias{bhist}
\title{Histogram with confidence intervals}
\usage{
bhist(x, b = NULL, z = 1.64, xlab = "", main = "")
}
\arguments{
\item{x}{a numerical vector}

\item{b}{the number of bins, or a vector of break locations.  If NULL,
chosen automatically by \code{\link{hist}}.}
}
\value{
None.
}
\description{
Same as \code{\link{hist}} except a confidence interval is drawn around each
bin height.
}
\details{
The width of the interval for height p is
\code{sqrt(p*(1-p)/n)*exp(-1/6/p/n)}.
}
\examples{

x <- c(rnorm(100,-2,0.5), rnorm(100,2,0.5))
b <- seq(-4,4,by=0.25)
bhist(x,b)

}
\author{
Tom Minka
}
