% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{break.quantile}
\alias{break.quantile}
\title{Equal-count breaks of a dataset}
\usage{
break.quantile(x, n = 2, plot = F, pretty = F)
}
\description{
Computes a set of breaks using the equal-count algorithm.
}
\author{
Tom Minka
}
