% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{break.ts}
\alias{break.ts}
\title{Change-point analysis by clustering}
\usage{
break.ts(x, n = 2, trace = T, same.var = T, ...)
}
\arguments{
\item{x}{a numerical vector or \code{ts} object}

\item{n}{the desired number of segments}

\item{trace}{If TRUE, shows a merging trace via
\code{\link{plot.hclust.trace}}}

\item{same.var}{argument passed to \code{\link{ward}}}
}
\value{
A vector of time breaks.  The breaks are also plotted visually via
\code{\link{plot.segments.ts}}.
}
\description{
Divide a time-series into homogenous segments.
}
\details{
Calls \code{\link{ward}} with \code{sortx=F} to cluster the series into
segments.  Only the marginal distribution of data is used; temporal
smoothness, for example, is ignored.
}
\examples{

data(LakeHuron)
# single major change
break.ts(LakeHuron,2)
# merging trace suggests n=6 is also interesting:
break.ts(LakeHuron,6)
# interesting oscillation

data(treering)
break.ts(treering[1:500],9,same=T)
break.ts(treering[1:100],7,same=T)
# interesting multiscale structure

x <- c(rnorm(100),rnorm(300)*3,rnorm(200)*2)
b <- break.ts(x,3,same=F)
plot(x,type="l")
plot.breaks(b)

}
\seealso{
\code{\link{plot.segments.ts}}, \code{\link{plot.breaks}}
}
\author{
Tom Minka
}
