% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamining-package.R
\name{Cars}
\alias{Cars}
\alias{Housing}
\alias{States}
\alias{va.deaths}
\title{Cars, Housing, States, and VA Deaths}
\value{
The specified dataset is loaded as a \code{\link{data.frame}}.  The
first three also define a transformed dataset (ending in `T') where the
variables are standardized.
}
\description{
Cleaned-up versions of some popular R datasets.
}
\details{
The variables are given more informative names, some irrelevant variables
are dropped, erroneous values are fixed, and some variables are transformed
(in the standardized case).
}
\examples{

data(States)
dev.new()
hist(States)
hist(StatesT)
w <- pca(StatesT, 2)
text_plot(project(StatesT, w), asp=1, cex=.6)
plot.axes(w)

}
\seealso{
\code{\link{Cars93}},\code{\link{Boston}},\code{\link{state}},\code{\link{VADeaths}}
}
\author{
Tom Minka
}
