% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{color.key}
\alias{color.key}
\title{Add a thin key to the top of a plot}
\usage{
color.key(
  col = NULL,
  pch = NULL,
  labels = NULL,
  breaks = NULL,
  digits = 2,
  cex = 0.75 * par("cex")
)
}
\arguments{
\item{col}{a vector of colors, one for each group.  Assumed black if none
given.}

\item{pch}{a vector of symbols, one for each group.  If less than two
distinct symbols are given, no symbols are put in the key.}

\item{labels}{a character vector giving the label of each group.}

\item{breaks}{a numeric vector of boundaries between groups.}

\item{digits}{number of digits to use for \code{breaks}.}

\item{cex}{character expansion factor.}
}
\value{
A thin bar is placed at the top of the plot, just inside the
plotting area.  The bar is divided into equal-sized lengths, colored
according to \code{col}.  If \code{pch} is given, that symbol is plotted in
the center of each segment.  If \code{labels} is given, each label is placed
at the center of the corresponding segment, just above the plotting area
(using \code{\link{mtext}}).  If \code{breaks} is given, the boundary
between segments is also labeled with a number.
}
\description{
Similar to \code{\link{legend}} but takes up less space.
}
\examples{

data(iris)
y = as.numeric(iris$Species)
plot(Sepal.Width ~ Sepal.Length, iris,col=y,pch=y)
color.key(1:3,1:3,levels(iris$Species))

}
\seealso{
\code{\link{color.plot}}
}
\author{
Tom Minka
}
