% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{color.plot.loess}
\alias{color.plot.loess}
\title{Contour plot of a regression surface}
\usage{
\method{color.plot}{loess}(
  object,
  x,
  res = 50,
  fill = F,
  add = fill,
  clip = T,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{\link{loess}} object.}

\item{res}{resolution of the sampling grid in each direction.}

\item{fill}{passed to \code{\link{contour.plot}}.}

\item{add}{If \code{TRUE}, the contours are added to an existing plot.
Otherwise a new plot is created.}

\item{clip}{a polygon over which the surface is to be defined.  Possible
values are \code{FALSE} (no clipping), \code{TRUE} (clip to the convex hull
of the data), or a matrix with two columns specifying (x,y) coordinates.}

\item{...}{extra arguments to \code{\link{color.plot.data.frame}}}

\item{data}{data to use instead of \code{model.frame(object)}.}
}
\value{
A plot is produced.
}
\description{
The regression surface is evaluated at all points on a grid, clipped values
are set to \code{NA}, and \code{contour.plot} is used to plot the contours.
}
\details{
If \code{add=FALSE}, the data is plotted on top using
\code{color.plot.data.frame}.
}
\examples{

data(Housing)
fit = loess(Price ~ Rooms + Low.Status, Housing)
color.plot(fit)

}
\seealso{
\code{\link{contour.plot}}
}
\author{
Tom Minka
}
