% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{color.plot}
\alias{color.plot}
\alias{color.plot.default}
\title{Plot subgroups as colored points}
\usage{
color.plot(object, ...)
}
\arguments{
\item{x, y}{numeric vectors of the same length.}

\item{z}{a numeric vector or factor, same length as \code{x} and \code{y}.}

\item{labels}{a character vector of labels, same length as \code{z}.  If
NULL, cases are plotted as points (default).}

\item{axes}{If FALSE, no axes are plotted (but there may still be a key).}

\item{key}{If FALSE, no key is plotted.}

\item{add}{If FALSE, a new plot is made.  If TRUE, points or labels are
added to the existing plot.}

\item{nlevels}{an integer.  If \code{z} is numeric, it is color-coded using
this many levels.  (If \code{z} is a factor, color-coding follows the factor
levels.)}

\item{color.palette}{a vector of colors, arbitrary length, or a function
with integer argument which generates a vector of colors (e.g.
\code{\link{YlGnBu.colors}}).  Used if \code{col} is not specified.  If
shorter than the number of levels, colors will be recycled and the plotting
symbol will change.}

\item{col}{a vector of colors, as in a call to \code{\link{plot}}.  Used to
explicitly set the color of each point.}

\item{pch.palette}{a vector of plotting symbols, arbitrary length.  If
\code{labels=NULL}, the plot symbol will rotate through these when there
aren't enough colors.}

\item{pch}{a vector of plotting symbols, as in a call to \code{\link{plot}}.
Used to explicitly set the symbol of each point.}

\item{digits}{the number of digits to use in the color key when \code{z} is
numeric.}

\item{bg}{the background color}

\item{mar}{figure margins (see \code{par}).  If not specified, appropriate
margins will be chosen automatically, which will not necessarily match the
current value of \code{par("mar")}.}
}
\value{
A plot is produced.
}
\description{
Like \code{\link{plot}} and \code{\link{text.plot}} but colors according to
a third variable.
}
\details{
Each (x,y) point is plotted with a color determined by \code{z}.  If
\code{z} is a factor, each factor level is in a different color.  If
\code{z} is numeric, then it is color-coded by assigning each quantile a
different color.

Despite the name, this function can also make plots using different symbols
instead of colors.  For example, if \code{color.palette=1} then all points
will be black but use different symbols.

When \code{labels != NULL}, the result is equivalent to
\code{\link{text.plot}} with colors.

If \code{key=TRUE} and there is more than one color or symbol on the plot, a
key is displayed at the top of the figure.
}
\note{
This function sets the figure margins permanently, so that you can
draw on the color plot.  Unfortunately, this also means future plots will
use the same margins, until you change them with \code{par("mar")}.
}
\examples{

# See the examples for color.plot.data.frame

}
\seealso{
\code{\link{color.plot.data.frame}},\code{\link{color.plot.loess}},\code{\link{color.plot.glm}},\code{\link{color.plot.knn}},\code{\link{color.plot.tree}},\code{\link{YlGnBu.colors}}
}
\author{
Tom Minka
}
