% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{contour.plot}
\alias{contour.plot}
\title{Display contours}
\usage{
\method{contour}{plot}(
  x,
  y,
  z,
  fill = F,
  levels = NULL,
  nlevels = if (is.null(levels)) 4 else length(levels),
  level.data = z,
  zlim,
  equal = F,
  pretty = T,
  lwd = 2,
  drawlabels = F,
  color.palette = YlGnBu.colors,
  bg = grey(0.65),
  key = T,
  zlab = "",
  ...
)
}
\arguments{
\item{x, y}{locations at which the values in \code{z} are measured.}

\item{z}{a matrix containing the values to be plotted (NAs are allowed).}

\item{fill}{If \code{TRUE}, makes a filled contour plot.}

\item{levels}{numeric vector of levels at which to draw contour lines.  The
next few arguments only apply to the case when \code{levels==NULL}.}

\item{nlevels}{The number of contour lines to draw.}

\item{level.data}{numeric vector to use for computing \code{levels}.}

\item{zlim}{The range that the levels should cover.  Defaults to the range
of \code{level.data}.}

\item{equal}{If \code{TRUE}, the levels are equally spaced over \code{zlim}.
Otherwise they match the quantiles of \code{level.data}.}

\item{pretty}{If \code{TRUE}, the levels are chosen to be round numbers.}

\item{lwd}{width of the contour lines.}

\item{drawlabels}{If \code{TRUE}, the contour lines are labeled with their
level.}

\item{color.palette}{a vector of colors, or a function which takes a number
and returns a vector of that length.}

\item{bg}{the background color.}

\item{key}{If \code{TRUE}, a color key is drawn at the top of the plot.  If
\code{key==2}, the key of \code{filled.contour} is used.}

\item{...}{extra arguments passed to \code{contour} or
\code{filled.contour}.}

\item{main}{the title of the plot.}
}
\description{
Create a filled or unfilled contour plot.
}
\details{
This is a wrapper function for \code{\link{contour}} and
\code{\link{filled.contour}} whose main purpose is to provide a uniform
interface and provide a decent automatic choice of levels.

If \code{equal=FALSE}, the levels are chosen according to the equal-count
algorithm of \code{\link{break.quantile}}.
}
\examples{

# see the examples for color.plot.loess

}
\seealso{
\code{\link{contour}},\code{\link{filled.contour}},
\code{\link{color.plot.loess}}
}
\author{
Tom Minka
}
