% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{data_image}
\alias{data_image}
\title{Data Image}
\usage{
data_image(x, las = 1, reorder = T, scale = T, ...)
}
\arguments{
\item{x}{a matrix or data frame.}

\item{reorder}{If \code{TRUE}, the rows and columns are reordered so that
the data along each row and each column follows a linear trend.  Can also be
a vector of two logicals, indicating separately whether to reorder rows
and/or columns.}

\item{scale}{If \code{TRUE}, each column is scaled to have minimum 0 and
maximum 1.}

\item{...}{extra arguments to \code{\link{image.default}}.}

\item{col}{a vector of colors to use for representing values.}
}
\description{
Each value in a data matrix is represented by a colored pixel.
}
\examples{

data(USJudgeRatings)
data_image(USJudgeRatings,col=RC.colors(32))

}
\references{
M. C. Minnotte and R. W. West. "The Data Image: A Tool For
Exploring High Dimensional Data Sets." Proceedings of the ASA Section on
Statistical Graphics, 1998. \url{http://citeseer.nj.nec.com/72085.html}
\url{http://math.usu.edu/~minnotte/research/pubs.html}
}
\seealso{
\code{\link{YR.colors}}
}
\author{
Tom Minka
}
