% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMcorMatrix_function.R
\name{DMcorMatrix}
\alias{DMcorMatrix}
\title{A Correlation Matrix Function}
\usage{
DMcorMatrix(data, lower, upper, cut_off)
}
\arguments{
\item{Data, }{lower, upper, cut_off}
}
\description{
This function allows you to construct a correlation Matrix
}
\examples{
x=rnorm(100,5)
y=rnorm(100,5)
z=rnorm(100,5)
d=data.frame(x,y,z)
DMcorMatrix(data=d, lower=1, upper=3, cut_off=0)
}
\keyword{Correlation}
\keyword{matrix,}
\keyword{variable}
