% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{hist.hclust}
\alias{hist.hclust}
\title{Histogram with hierarchical cluster breaks}
\usage{
\method{hist}{hclust}(hc, x, k = 2:5)
}
\arguments{
\item{x}{the numerical vector that was clustered to produce \code{h}}

\item{k}{a vector of the cluster cardinalities to plot}

\item{h}{an \code{hclust} object}
}
\value{
A histogram of \code{x} is shown with blue lines cutting the x-axis.
The tallest lines correspond to divisions made at the top of the hierarchy.
By reading top to bottom, you can see how each cluster is subdivided. This
can be far more illuminating than a plot of the hierarchy as a tree.
}
\description{
A representation of a one-dimensional hierarchical clustering
}
\seealso{
\code{\link{boxplot.hclust}}, \code{\link{ward}},
\code{\link{break_ward}}
}
\author{
Tom Minka
}
