% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{labeled_curves}
\alias{labeled_curves}
\title{Plot labeled curves}
\usage{
labeled_curves(
  x,
  y,
  se = NULL,
  labels,
  xtick = names(x),
  xlab = NULL,
  ylab,
  type = "o",
  group,
  color.palette = default.colors(6),
  col,
  lty.palette = 1:6,
  lty,
  lwd = 2,
  jitter = 0,
  legend. = NULL,
  move = T,
  dryrun = F,
  cex = 1,
  horizontal = F,
  mar,
  bg = par("bg"),
  ylim,
  main = "",
  xaxt = par("xaxt"),
  yaxt = par("yaxt"),
  ...
)
}
\arguments{
\item{x}{a numeric vector giving the horizontal position of each point.}

\item{y}{a matrix or data frame giving the vertical position of each point.
Each column defines one curve.}

\item{se}{a matrix or data frame, the same size as \code{y}, giving an error
bound on each value.}

\item{labels}{a character vector of labels for the lines.  Defaults to the
column names of \code{y}.}

\item{xlab, ylab}{axis labels, by default the dimnames of \code{y}.}

\item{type}{indicates whether to draw points, lines, or both.  See
\code{\link{plot}}.}

\item{group}{a numeric vector specifying how to group cases when assigning
colors.}

\item{color.palette}{a vector of colors, arbitrary length, or a function
with integer argument which generates a vector of colors.  Used if
\code{col} is not specified.  If shorter than the number of curves, colors
will be recycled and the line style will change.}

\item{col}{a vector of colors, as in a call to \code{\link{plot}}.  Used to
explicitly set the color of each curve.}

\item{lty.palette}{a vector of line styles, arbitrary length.  The line
style will rotate through these when there aren't enough colors.}

\item{lty}{a vector of line styles, as in a call to \code{\link{plot}}.
Used to explicitly set the style of each curve.}

\item{lwd}{line width.}

\item{jitter}{the amount by which to jitter the error bars, to avoid
overplotting.  \code{jitter=0.02} is usually sufficient.}

\item{cex}{character expansion factor.}

\item{horizontal}{If \code{TRUE}, the axes are laid out horizontally, so
that the curves run vertically.}

\item{mar}{a numerical vector giving the lines of margin on the four sides
of the plot (see \code{\link{par}}).  If missing, it is set according to
\code{\link{auto.mar}}.}

\item{bg}{background color.}

\item{ylim}{desired plotting limits.}

\item{main}{title for the plot.}

\item{...}{extra arguments for low-level plotting commands.}

\item{legend}{If \code{TRUE}, the labels are placed in a legend.  Otherwise
the labels are placed next to the lines.}
}
\description{
Draws multiple curves, each with a properly-placed label and unique
color/linestyle.
}
\details{
Point \code{j} in curve \code{i} is at \code{(x[j],y[j,i])}.  Thus all
curves must be the same length, and have points at the same horizontal
positions.  If \code{y[i,j]=NA}, then the curve has a break at the previous
point, and resumes at the next \code{non-NA} point.

If \code{se} is given, then an error bar will be placed around each point.
}
\seealso{
\code{\link{parallel_plot}}
}
\author{
Tom Minka
}
