% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{make.unique}
\alias{make.unique}
\title{Make character strings unique}
\usage{
make.unique(names, sep = ".", start = 2)
}
\arguments{
\item{names}{a character vector}

\item{sep}{a character string used to separate a duplicate name from its
sequence number.}
}
\value{
a character vector of same length as \code{names} with duplicates
changed.
}
\description{
Makes the elements of a character vector unique by appending sequence
numbers to duplicates.
}
\details{
The algorithm used by \code{make.unique} has the property that
\code{make.unique(c(A,B)) = make.unique(c(make.unique(A),B))}.

In other words, you can append one string at a time to a vector, making it
unique each time, and get the same result as applying \code{make.unique} to
all of the strings at once.

If character vector \code{A} is already unique, then
\code{make.unique(c(A,B))} preserves \code{A}.
}
\examples{

make.unique(c("a","a","a"))
make.unique(c(make.unique(c("a","a")),"a"))

make.unique(c("a","a","a.2","a"))
make.unique(c(make.unique(c("a","a")),"a.2","a"))

rbind(data.frame(x=1),data.frame(x=2),data.frame(x=3))
rbind(rbind(data.frame(x=1),data.frame(x=2)),data.frame(x=3))

}
\seealso{
\code{\link{make.names}}
}
\author{
Thomas P Minka
}
\keyword{character}
