% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{make_numeric_data_frame}
\alias{make_numeric_data_frame}
\title{Convert data frame columns to numeric}
\usage{
make_numeric_data_frame(data, exclude = NULL, warn = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{exclude}{columns to exclude}

\item{warn}{allow warnings}
}
\description{
Factors are also converted to numeric.
}
\examples{
make_numeric_data_frame(iris)
}
