% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{merge.factor}
\alias{merge.factor}
\title{Merge factor levels}
\usage{
\method{merge}{factor}(f, x, n, same.var = T, trace = T, xlab = NA, ylab = NA)
}
\arguments{
\item{f}{a factor}

\item{x}{a numerical vector, same length as \code{f}}

\item{n}{the desired number of factor levels}

\item{same.var}{argument passed to \code{\link{ward}}}

\item{trace}{If TRUE, a merging trace is plotted
(\code{\link{plot.hclust.trace}})}

\item{xlab, ylab}{axis labels.  If NA, taken from f and x arguments.}
}
\value{
A new factor, same length as \code{f}, but with \code{n} levels.
}
\description{
Merges factor levels with similar response distributions (assumed normal).
}
\details{
Calls \code{ward(split(x,f))} to get a tree, cuts the tree, and constructs a
new factor.  The tree is shown via \code{\link{boxplot.hclust}}.
}
\examples{

n <- 20
x <- c(rnorm(n)+1, rnorm(n)+2, rnorm(n)*4+2)
f <- gl(3,n)
levels(f) <- c("a","b","c")
merge.factor(f,x,2,same.var=T)
merge.factor(f,x,2,same.var=F)

# an ordered factor
data(va.deaths)
merge.factor(va.deaths$Age,va.deaths$Rate,2)

}
\author{
Tom Minka
}
