% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{merge.hist}
\alias{merge.hist}
\title{Merge histogram bins}
\usage{
\method{merge}{hist}(x, b = NULL, n = b, trace = T)
}
\arguments{
\item{x}{a numerical vector}

\item{b}{the starting number of bins, or a vector of starting break
locations. If NULL, chosen automatically by \code{\link{hist}}.}

\item{n}{the desired number of bins.}
}
\value{
A vector of bin breaks, suitable for use in \code{\link{hist}},
\code{\link{bhist}}, or \code{\link{cut}}. Two plots are shown: a
\code{\link{bhist}} using the returned bin breaks, and a merging trace.  The
trace shows, for each merge, the chi-square distance of the bins which were
merged.  This is useful for determining the appropriate number of bins.  An
interesting number of bins is one that directly precedes a sudden jump in
the chi-square distance.
}
\description{
Quantize a variable by merging similar histogram bins.
}
\details{
The desired number of bins is achieved by successively merging the two most
similar histogram bins.  The distance between bins of height (f1,f2) and
width (w1,w2) is measured according to the chi-square statistic
\deqn{w1*(f1-f)^2/f + w2*(f2-f)^2/f} where f is the height of the merged
bin: \deqn{f = (f1*w1 + f2*w2)/(w1 + w2)}
}
\examples{

x <- c(rnorm(100,-2,0.5),rnorm(100,2,0.5))
b <- seq(-4,4,by=0.25)
merge.hist(x,b,10)
# according to the merging trace, n=5 and n=11 are most interesting.

x <- runif(1000)
b <- seq(0,1,by=0.05)
merge.hist(x,b,10)
# according to the merging trace, n=6 and n=9 are most interesting.
# because the data is uniform, there should only be one bin,
# but chance deviations in density prevent this.
# a multiple comparisons correction in merge.hist may fix this.

}
\author{
Tom Minka
}
