% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{merge.table.cost}
\alias{merge.table.cost}
\alias{merge.table.cost.inorder}
\alias{merge.table.cells}
\title{Internal routines}
\usage{
\method{merge}{table.cost}(x, v, cost = homogeneity)
}
\description{
Internal routines for \code{\link{merge.table}}
}
\author{
Tom Minka
}
