% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{mine.associations}
\alias{mine.associations}
\title{Associations in a contingency table}
\usage{
mine.associations(x, top = 10, targets = NULL, z = 1)
}
\arguments{
\item{x}{a contingency table}

\item{top}{the number of top associations to return}

\item{targets}{a character vector of table variables, one of which must be
included in every returned association.}
}
\value{
A data frame where the first column is \code{Lift} and the remaining
columns correspond to the variables of \code{x}. Each row describes a cell
of \code{x}, where marginalized variables have value \code{NA}. The lift
value describes the ratio of the actual count in the cell versus the
expected count under independence.
}
\description{
Find unusually frequent variable combinations.
}
\details{
Enumerates all two-way marginal tables, sorts the cells by lift, and returns
the top.  The formula for lift is \deqn{\frac{p(i,j)}{p(i)p(j)}}
}
\examples{

data(Titanic)
mine.associations(Titanic)
# Females are twice as likely to survive as the average person.
# Members of 3rd class are twice as likely to be children as the average person.
# Etc.

# focus on associations with survival
mine.associations(Titanic,target="Survived")

# focus on children
mine.associations(Titanic[,,1,],target="Survived")

}
\author{
Tom Minka
}
