% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{parallel_plot}
\alias{parallel_plot}
\title{Parallel-coordinate plot}
\usage{
parallel_plot(
  x,
  yscale = c("linear", "range", "none"),
  flipy = F,
  xscale = c("equal", "linear", "none"),
  proto = T,
  flipx = F,
  yaxt = if (yscale == "none") "s" else "n",
  type = 1,
  ...
)
}
\arguments{
\item{x}{a data matrix or frame.  Only numeric columns will be
plotted.}

\item{yscale}{describes how the columns should be scaled.}

\item{xscale}{describes how the columns should be placed.}

\item{proto}{If \code{TRUE} and the response variable is a factor,
then only the mean value for each response is plotted.
See \code{\link{prototypes}}.}

\item{...Extra}{arguments to \code{\link{labeled.curves}}.}
}
\description{
Plots a data matrix using parallel coordinates.
}
\details{
Let the rows of the matrix be \emph{cases} and the columns be
\emph{dimensions}.
Each dimension is laid out as a vertical axis, and the value for each
case is plotted as a point on that axis.  Points corresponding to the
same case are connected by lines.  Thus each case is represented by a
curve or \emph{profile}, moving horizontally from dimension to dimension.

To enhance readability, the dimensions should usually be shifted to have
a common center and scaled to have
comparable units.  This is controlled by \code{yscale}.
If \code{yscale="linear"}, the dimensions are automatically shifted
and scaled
by the linear profiles method.  That is, the profiles are
made to be as straight as possible.
This may result in negative scaling for a dimension,
i.e. the axis is reversed so that moving up means a lower value.
When this happens, the dimension name is prepended with a minus sign.
If \code{yscale="range"}, the dimensions are shifted and scaled (by a
positive number) so that the data ranges from 0 to 1 on each axis.
If \code{yscale="none"}, the dimensions are plotted as-is.

Another important choice is the ordering and spacing of the axes
on the plot.
If \code{xscale="linear"}, the axes are placed according to the
linear profiles method.
If \code{xscale="equal"}, the axes are placed similarly to
\code{"linear"} but with the constraint that they must be equally
spaced.
If \code{xscale="none"}, the axes are placed in the order
that they appear in the matrix.

If the data frame has a categorical response, the profiles are colored
according to the response.  The color scheme can be adjusted using the
arguments to \code{\link{labeled.curves}}.
}
\examples{
data(iris)
parallel_plot(iris)
parallel_plot(iris,proto=F,labels=NULL)
}
\references{
A. Inselberg and B. Dimsdale.
"Parallel coordinates: A tool for visualizing multidimensional geometry."
Proc. of Visualization 90, p. 361-78, 1990.

E. J. Wegman.
"Hyperdimensional Data Analysis Using Parallel Coordinates."
JASA 85:664-675, 1990.
}
\seealso{
\code{\link{parallel.cases}},\code{\link{star.plot}}
}
\author{
Tom Minka
}
