% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{plot.hclust.trace}
\alias{plot.hclust.trace}
\title{Plot a merging trace}
\usage{
\method{plot}{hclust.trace}(h, k = 1:10)
}
\arguments{
\item{h}{an \code{hclust} object}

\item{ka}{vector of the cluster cardinalities to plot}
}
\value{
The trace shows, for each merge, the
  distance of the clusters which were merged.  This is useful for determining
  the appropriate number of clusters.  An interesting number of clusters is one
  that directly precedes a sudden jump in distance.
}
\description{
Plots the cost of successive merges in a hierarchical clustering
}
\seealso{
\code{\link{ward}}, \code{\link{break_ward}}
}
\author{
Tom Minka
}
