% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{plot.segments.ts}
\alias{plot.segments.ts}
\title{Plot time-series segments}
\usage{
\method{plot}{segments.ts}(x, b, lwd = 2, ...)
}
\arguments{
\item{x}{a numeric vector or \code{ts} object}

\item{b}{a numeric vector of break times}
}
\value{
Plots the time series as a line, then draws a blue horizontal line
  through each segment, at the mean value of the segment.
}
\description{
Shows how the mean level of a time-series varies between
  segments.
}
\seealso{
\code{\link{break.ts}},
  \code{\link{plot_breaks}}
}
\author{
Tom Minka
}
