% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{plot_breaks}
\alias{plot_breaks}
\title{Draw break locations}
\usage{
plot_breaks(b)
}
\arguments{
\item{b}{a numeric vector}
}
\value{
Uses \code{\link{segments}} to draw thick blue vertical lines on top
  of the current plot, at the given locations.
}
\description{
Draws lines to indicate break locations.
}
\seealso{
\code{\link{plot.hclust.breaks}},
  \code{\link{plot.segments.ts}},
  \code{\link{break.kmeans}}
}
\author{
Tom Minka
}
