% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{predict_plot.lm}
\alias{predict_plot.lm}
\title{Plot predictors versus residuals.}
\usage{
\method{predict_plot}{lm}(object, data, partial = F, ylab = NULL, ...)
}
\arguments{
\item{object}{the output of \code{lm}.}

\item{data}{a data frame to use instead of \code{model.frame(object)}.}

\item{partial}{If \code{TRUE}, plot partial residuals instead of residuals.}

\item{ylab}{axis label.}

\item{...}{extra arguments for \code{\link{predict_plot.data.frame}}.}
}
\value{
A plot similar to \code{\link{predict_plot}}, but where the vertical
  axis is residuals.  These plots can be used to judge which predictors
  should be added to the model.
}
\description{
Makes a matrix of pairwise scatterplots with lowess-type trend lines.
}
\details{
Partial residuals are computed by fitting a new model with the
predictor removed, which is different from \code{residuals(type="partial")}.
}
\examples{
# see the examples for predict_plot
}
\seealso{
\code{\link{predict_plot}}
}
\author{
Tom Minka
}
