% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{predict_plot}
\alias{predict_plot}
\alias{predict_plot.data.frame}
\alias{predict_plot.formula}
\title{Plot predictors versus response}
\usage{
predict_plot(...)
}
\arguments{
\item{...}{extra arguments passed to \code{predict_plot.data.frame} or
\code{\link{plot}}.}

\item{formula}{a formula specifying the response and predictor variables}

\item{data, x}{a data frame with at least two columns}

\item{partial}{a model from which to compute partial residuals (used
by \code{\link{predict_plot.lm}}).}

\item{mcol, mlwd}{If plotting partial residuals of an \code{lm},
the color and width of the model predictions.}

\item{layout}{a vector \code{c(rows,cols)} specifying the desired
layout of panels.  Otherwise chosen automatically based on the size
of the plotting window.}

\item{highlight}{a logical vector specifying which predictors to highlight.}

\item{se}{If \code{TRUE}, show standard errors in linecharts.}

\item{scol, slwd}{color and width of trend lines.}

\item{span, degree, family}{parameters for the trend line (see \code{loess}).}

\item{rtype}{how a factor response should be handled when drawing a
trend line.}

\item{identify.pred}{A character vector of predictor names for which
to interactively \code{\link{identify}} points.  If \code{TRUE},
done for all predictors.}

\item{mar}{margins within each panel}

\item{xaxt, yaxt}{arguments to \code{par}}

\item{col}{plotting color for symbols}

\item{asp}{Aspect ratio for each panel.  If \code{"auto"}, the aspect
ratio is chosen automatically based on the trend line and
\code{\link{auto.aspect}}.}

\item{given, given.lab, nlevels, pretty, key, bg, color.palette, pch.palette}{used for conditioning plots.}

\item{main, xlab, ylab}{axis labels.}
}
\value{
If the predictor is numeric, makes a scatterplot with loess line on top.
  If the predictor is a factor, makes a \code{\link{linechart}}.
}
\description{
Makes a matrix of pairwise scatterplots with lowess-type trend lines.
}
\examples{
data(Cars)
predict_plot(Price ~ ., CarsT)
fit <- lm(Price ~ ., CarsT)
predict_plot(Price~ ., CarsT, partial=fit)
# same thing using predict_plot.lm
predict_plot(fit, partial = TRUE)
}
\seealso{
\code{\link{loess}}, \code{\link{model.plot}}
}
\author{
{Tom Minka}
}
