% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{predictor_vars}
\alias{predictor_vars}
\title{Variables out of which the predictors are constructed}
\usage{
predictor_vars(object)
}
\arguments{
\item{object}{that inherits from a data.frame or terms object}
}
\value{
vector of predicts
}
\description{
If the object inherits from a \code{data.frame} then it is assumed the
response is in the first column.
}
\examples{
fit <- lm(mpg ~ .^2, data=mtcars)
predictor_vars(fit)
predictor_vars(mtcars)
}
