% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project data into fewer dimensions}
\usage{
project(x, w)
}
\arguments{
\item{x}{a data frame}

\item{w}{a matrix with named rows and columns}
}
\value{
A data frame where the variables named in the rows of \code{w} are
  replaced by new variables named in the columns of \code{w}.
  Other variables are left unchanged.
}
\description{
Reduces the dimensionality of a data set.
}
\details{
Each column of \code{w} specifies a new variable, which is to be
  constructed by combining existing variables according to the given weights.
}
\examples{
data(iris)
w = projection(iris, k=2)
# w only involves the continuous attributes
# the new variables are h1 and h2
x = project(iris, w)
# in RStudio use dev.new()
color.plot(x)
plot.axes(w)
}
\seealso{
\code{\link{pca}}, \code{\link{projection}}, \code{\link{plot.axes}}
}
\author{
Tom Minka
}
