% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_array}
\alias{read_array}
\alias{write_array}
\title{Read and write arrays}
\usage{
read_array(file, sep = "", quote = "\\"", skip = 0)
}
\arguments{
\item{file}{either a character string naming a file or a connection which the data are to be read from or written to. "" indicates input from the console for reading and output to the console for writing.}

\item{sep}{the field separator string. Values on each line of the file are separated by this string.}

\item{quote}{the set of quoting characters as a single character string.}

\item{skip}{the number of lines of the input file to skip before beginning to read data.}
}
\value{
An array.
}
\description{
Read data into an array from the console or file
}
\details{
The array can have one of several formats.  The preferred format,
produced by \code{\link{write.array}}, looks like \code{
    cvar.nam
    rvar.1.nam   ... rvar.k.nam    cvar.lev.1 ... cvar.lev.l
    rvar.1.lev.1 ... rvar.k.lev.1  ...        ... ...
  }

Modified from rw1030/library/base/R/base/read.ftable
}
\examples{
data(HairEyeColor)
fn <- tempfile()
write_array(HairEyeColor, fn)
read_array(fn)
}
\author{
Tom Minka
}
