% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{response.var}
\alias{response.var}
\title{TODO: remove this
Get response variable}
\usage{
response.var(object)
}
\arguments{
\item{object}{formula, model frame, or fitted model}
}
\value{
the name of the response variable
}
\description{
This works for
}
\examples{
response.var(a ~ b + c)
response.var(~ b + c) # gives '.'
response.var(Height + Volume ~ Girth) # incorrect
response.var(log(Volume) ~ Girth) # incorrect

}
