% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{response_var}
\alias{response_var}
\title{Get response variable}
\usage{
response_var(obj)
}
\arguments{
\item{obj}{formula, model frame, or fitted model}
}
\value{
character string
}
\description{
Get response variable
}
\examples{
response_var(a ~ b + c)
# response_var(~ b + c) # gives error
response_var(Height + Volume ~ Girth)
response_var(log(Volume) ~ Girth)
}
