% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{sort.data.frame}
\alias{sort.data.frame}
\title{Sort rows of a data frame}
\usage{
\method{sort}{data.frame}(x, decreasing = FALSE, f = ncol(x), ...)
}
\arguments{
\item{x}{a data frame}

\item{f}{the name or number of a column of \code{x}, or a numeric vector
to play the role of a column of \code{x}. Default: Last column.}
}
\value{
A reordered version of \code{x} where the values in column \code{f} are in
ascending order.
}
\description{
Re-orders data frame rows so that a column is sorted.
}
\examples{
data(mtcars)
sort(mtcars, f = "mpg")
}
\seealso{
\code{\link{sort_cells}}
}
\author{
Tom Minka
}
