% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{sort_cells}
\alias{sort_cells}
\title{Sort an array}
\usage{
sort_cells(x)
}
\arguments{
\item{x}{an array}
}
\value{
A data frame representation of \code{x}, sorted by cell value.
}
\description{
Sort the cells of an array
}
\details{
Converts \code{x} to a data frame and calls
  \code{\link{sort.data.frame}}.
}
\examples{
data(Titanic)
sort_cells(Titanic)
data(HairEyeColor)
sort_cells(HairEyeColor)
}
\seealso{
\code{\link{sort.data.frame}}
}
\author{
Tom Minka
}
