% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb.R
\name{sort_levels}
\alias{sort_levels}
\title{Sort levels of a factor}
\usage{
sort_levels(f, x, fun = median)
}
\arguments{
\item{f}{a factor}

\item{x}{a vector, same length as \code{f}}

\item{fun}{function to use to summarize groups of \code{x}}
}
\value{
A copy of \code{f} with re-ordered levels.
}
\description{
The levels of a factor are sorted according to a summary statistic.
}
\examples{
data(OrchardSprays)
sort_levels(OrchardSprays$treatment, OrchardSprays$decrease)
sort_levels(OrchardSprays$treatment, OrchardSprays$decrease, fun=mean)
}
\author{
Tom Minka
}
