% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{sort_table}
\alias{sort_table}
\title{Sort rows and columns of a contingency table}
\usage{
sort_table(x, k = NULL)
}
\arguments{
\item{x}{a contingency table}
}
\description{
Score rows and columns via correspondence analysis, then sort
the scores.
}
\details{
Only for 2D tables right now.
}
\examples{
data(blood)
sort_table(blood)

}
\seealso{
\code{\link{mosaicplot}}
}
