% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{star_plot}
\alias{star_plot}
\title{Reordered and scaled star plot}
\usage{
star_plot(x, draw.segments = T, proto = T, reorder = T, scale = T, ...)
}
\arguments{
\item{x}{a matrix or data frame.  Each row makes one star.}

\item{proto}{If \code{TRUE} and the response variable is a factor,
then only the mean value for each response is plotted.
See \code{\link{prototypes}}.}

\item{reorder}{If \code{TRUE}, the rows and columns are reordered so
that the data along each row and each column follows a linear trend.
Can also be a vector of two logicals, indicating separately
whether to reorder rows and/or columns.}

\item{scale}{If \code{TRUE}, each column is scaled to have minimum
0 and maximum 1.}

\item{...}{additional arguments for \code{\link{stars}}.}
}
\description{
Automatically reorders and scales the variables to make a
readable star plot.
}
\examples{
data(iris)
star_plot(iris)
}
\seealso{
\code{\link{stars}}
}
\author{
{Tom Minka}
}
