% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamining-package.R
\name{state.pop}
\alias{state.pop}
\title{U.S. state population 1790-1990}
\description{
U.S. state population 1790-1990
}
\details{
\code{state.pop} is a matrix containing the population of
  each U.S. state, including the District of Columbia,
  from 1790 to 1990 in ten-year increments.
  The rows are the years "1790" to "1990" and the columns are state names.
}
\examples{
  dotchart(sort(log(state.pop["1990",])))

  # distribution of growth rates over time
  rates <- diff(log(state.pop))
  boxplot(as.data.frame(t(rates)))
}
\references{
The data was obtained from
  \url{http://merrill.olm.net/pdocs/feas/pop/pop1790_1990/pii.txt}
  and edited into R format.
}
\seealso{
\code{\link{state.area}}, \code{\link{state.name}}, \code{\link{state}}
}
