% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{terms.data.frame}
\alias{terms.data.frame}
\title{does NOT throw an error if no terms

#' @export
terms <- function(...) UseMethod("terms")}
\usage{
\method{terms}{data.frame}(x, env = parent.frame(), ...)
}
\description{
#' @export
terms.default <- function(x,...) x$terms
}
