% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{text_plot}
\alias{text_plot}
\alias{text.plot}
\alias{text.plot.default}
\alias{text.plot.data.frame}
\alias{text.plot.formula}
\title{Make a plot of text labels}
\usage{
text_plot(object, ...)
}
\arguments{
\item{x, y}{numeric vectors, same length}

\item{data}{a data frame with at least two columns}

\item{formula}{a formula specifying a response and predictor variable}

\item{labels}{character vector, same length as \code{x}}

\item{srt}{rotation angle as in \code{\link{text}}.}

\item{adj}{text justification as in \code{\link{text}}.}
}
\description{
Like \code{\link{text}} except it creates a new plot with
limits chosen to make all labels visible.
}
\details{
The main job of \code{\link{text.plot}} is finding the right plot
limits so that all labels are visible.  It does this by computing the
bounding box of each label (taking into account text rotation and
justification) and solving a system of inequalities which ensure that
each label fits entirely into the plot window.
}
\examples{
data(state)
x <- data.frame(state.x77)
text_plot(x$Frost, x$HS.Grad, rownames(x))

# same thing, using text.plot.formula
text_plot(HS.Grad ~ Frost, x)

# notice how the limits change
text_plot(HS.Grad ~ Frost, x, srt=45)
text_plot(HS.Grad ~ Frost, x, srt=90)
}
\author{
Tom Minka
}
