\name{untitled1-package}
\alias{untitled1-package}
\alias{untitled1}
\docType{package}
\title{
\packageTitle{untitled1}
}
\description{
\packageDescription{untitled1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{untitled1}
\packageIndices{untitled1}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{untitled1}

Maintainer: \packageMaintainer{untitled1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
