% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{my.update}
\alias{my.update}
\title{#' @export
terms.table <- function(x,...) {
terms.data.frame(as.data.frame(x),...)
}}
\usage{
my.update(object, formula., ..., evaluate = TRUE)
}
\description{
#' @export
formula <- function(...) UseMethod("formula")
}
\section{minka: always return formula, not terms}{

}

